fix: prevent infinite loop for token trimming
This prevents the infinite loop of attempting to trim a message by introducing a maximum number of iterations that the trim function can attempt before giving up.

There is probably a better approach to also help with the summarizing/trimming of the tokens, but I figured just preventing the infinite loop was a good first step.

It also addresses some of the feedback around mixing characters and tokens.

Closes #3